[Github Gists] OAuth 2.0 support (#13403)
* feat(github-gists): Adds OAuth2.0 support and deprecates tokens

closes https://github.com/raycast/extensions/issues/13402

* refactor(github-gists): fixed lint issues

closes https://github.com/raycast/extensions/issues/13402

* refactor(github-gists): changed the readme screencast

closes https://github.com/raycast/extensions/issues/13402

* Update CHANGELOG.md and optimise images

---------
