+++ extensions/github-gist/CHANGELOG.md
@@ -1,5 +1,11 @@
 # GitHub Gist Changelog
 
+## [OAuth2.0 Support] - 2024-07-11
+
+- Uses Raycast OAuth2.0 GitHub integration with `repo read:user gist` scopes.
+- ⚠️Disables the personal access token preference. Users can delete their tokens after successful OAuth2.0 connection.
+- Minor refactoring and updated dependencies.
+
 ## [Refactor Command] - 2024-07-03
 
 - Simplify the code and improve performance
