+++ extensions/github-gist/README.md
@@ -2,29 +2,17 @@
 
 ### Getting Started
 
-This extension brings GitHub Gist support to Raycast through the personal access tokens. To get started, first:
+This extension brings GitHub Gist support to Raycast via OAuth 2.0. 
 
-- Login to your  GitHub instance (Eg. [https://github.com](https://github.com))
-- Click on your avatar image in the right upper corner
-- From the dropdown menu, click on `Settings`
-- On the left-hand side, click on `Developer settings`.
-- On the left-hand side, click on `Personal access tokens`.
-- On this page, we'll create a new access token. Click on `Generate new token` in the upper-right.
-- Leave a `note` for your token (Eg. `Raycast`). This will help you identify it in the future.
-- You'll need to check the following boxes to ensure this extension can perform properly:
-  - `repo`
-  - `gist`
-  - `user`
-- Click `Generate token` and save this value somewhere. **You'll only be able to see this once**.
+To get started, open any command and it will prompt you to connect to you GitHub instance via OAuth.
+You will be led to your browser where you might need to sign in to your [github](https://github.com).
+Once you accept the new/additional scopes for Raycast GitHub integration, the command will open back up 
+with results. **You'll only be able to see this once**.
+
+> If you used GitHub Personal Access Token for this extension, you can delete them after connecting through OAuth.
 
 > Future updates to this extension will offer more functionality that may require additional scopes be defined in this token.
 
 ### Screencast
 
-- Create Gist
-
-https://user-images.githubusercontent.com/36128970/159211152-65b06683-56a3-4db4-842f-3acb05c6ce62.mp4
-
-- Search Gist
-
-https://user-images.githubusercontent.com/36128970/159161962-70adae29-29c5-4026-82d0-19f23b8dcc22.mp4
+<img src="media/usage.gif" alt="Usage">
