+++ extensions/github-gist/src/util/utils.ts
@@ -1,7 +1,7 @@
 import { parse } from "path";
-import { GistItem } from "./gist-utils";
 import { Alert, confirmAlert, Icon } from "@raycast/api";
 import ActionStyle = Alert.ActionStyle;
+import { GistItem } from "./gist-utils";
 
 export const isEmpty = (string: string | null | undefined) => {
   return !(string != null && String(string).length > 0);
